try to factor in UVs when drawing tiles
Helps draw the numbers in the right place when tiles are partially
offscreen. Doesn't work perfectly but I suspect the game is just setting
UVs incorrectly in those cases, not much I can do about that. And it's
only a minor visual bug anyway, in fact some might like it.